% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFundingStage.R
\name{convertFundingStage}
\alias{convertFundingStage}
\title{Helper function to convert funding stage to human readable numbers}
\usage{
convertFundingStage(stage)
}
\value{
Human understandable format of the funding stage e.g. "M&A"
}
\description{
Helper function to convert funding stage to human readable numbers
}
\examples{
convertFundingStage("early_stage_venture")


}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
