% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFundingType.R
\name{convertFundingType}
\alias{convertFundingType}
\title{Helper function to convert funding types}
\usage{
convertFundingType(funding_type)
}
\value{
converted funding types
}
\description{
Helper function to convert funding types
}
\examples{
convertFundingType()

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
