% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertRevenueRange.R
\name{convertRevenueRange}
\alias{convertRevenueRange}
\title{Helper function to convert renvenue range to human readable numbers}
\usage{
convertRevenueRange(rev_range)
}
\value{
Human understandable format of the revenue range e.g. "$1M to $10M"
}
\description{
Helper function to convert renvenue range to human readable numbers
}
\examples{
convertRevenue("r_00001000")


}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
