% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTerms.R
\name{convertTerms}
\alias{convertTerms}
\title{Helper function to convert terms to human readable numbers}
\usage{
convertTerms(term)
}
\value{
Human understandable format of the terms e.g. "Cash & Stock"
}
\description{
Helper function to convert terms to human readable numbers
}
\examples{
convertTerms("cash")


}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
