% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currencyListParse.R
\name{currencyListParse}
\alias{currencyListParse}
\title{Helper function to parse a currency element}
\usage{
currencyListParse(fields_data, field)
}
\value{
parsed output
}
\description{
Helper function to parse a currency element
}
\examples{
currencyListParse(CurrencyElement)


}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
