% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicateTimeCheck.R
\name{duplicateTimeCheck}
\alias{duplicateTimeCheck}
\title{duplicateTimeCheck}
\usage{
duplicateTimeCheck(entities)
}
\arguments{
\item{entities}{the uuid/permalink(s) of the entity to lookup a card for}
}
\value{
a custom message that is printed to the console
}
\description{
Helper function to remove duplicates in the inputs (uuids/permalinks) of any lookupXXXCard() function and to inform the user of the approximate time the funciton will take to crunch the data.
}
\examples{
cardPaginationCheck(c(uuid1, uuid2))



}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
