% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entityIdCheck.R
\name{entityIdCheck}
\alias{entityIdCheck}
\title{Function to check and help out with uuid/permalink of an entity uuid to make it consistent}
\usage{
entityIdCheck(input)
}
\arguments{
\item{input}{Original input string as uuid/permalink to the function}
}
\value{
the string manipulated output
}
\description{
It will string manipulate the input to lower case,
trim whitespace at beginning and end,
and replace spaces with hyphens so that it can be passed in to an entity lookup function
}
\examples{
autocomplete(collection_ids = "categories", query = "Artificial Intelligence", limit = 5)

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
