% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAcquisitionStatus.R
\name{getAcquisitionStatus}
\alias{getAcquisitionStatus}
\title{Helper function to return the acquisition status dictionary}
\usage{
getAcquisitionStatus()
}
\value{
dictionary
}
\description{
Helper function to return the acquisition status dictionary
}
\examples{
getAcquisitionStatus()

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
