% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAcquisitionTypes.R
\name{getAcquisitionTypes}
\alias{getAcquisitionTypes}
\title{Helper function to return the acquisition types dictionary}
\usage{
getAcquisitionTypes()
}
\value{
dictionary
}
\description{
Helper function to return the acquisition types dictionary
}
\examples{
getAcquisitionTypes()

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
