% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCardsForPath.R
\name{getCardsForPath}
\alias{getCardsForPath}
\title{Helper function to return all cards for one of the paths}
\usage{
getCardsForPath(path, pretty_print = TRUE)
}
\arguments{
\item{path}{of interest}

\item{pretty_print}{TRUE by default, then it will print out an explanatory line before the list of available cards for a path}
}
\value{
all valid cards for a particular path
}
\description{
Helper function to return all cards for one of the paths
}
\examples{
getCardsFor("organizations")
getCardsFor("funding_rounds")

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
