% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectionIds.R
\name{getCollectionIds}
\alias{getCollectionIds}
\title{Helper function to return all colleciton ids}
\usage{
getCollectionIds()
}
\value{
all valid collection ids
}
\description{
Helper function to return all colleciton ids
}
\examples{
getCollectionIds()

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
