% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCompanyTypes.R
\name{getCompanyTypes}
\alias{getCompanyTypes}
\title{Helper function to return the company types dictionary}
\usage{
getCompanyTypes()
}
\value{
dictionary
}
\description{
Helper function to return the company types dictionary
}
\examples{
getCompanyTypes()

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
