% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEmployeeRanges.R
\name{getEmployeeRanges}
\alias{getEmployeeRanges}
\title{Helper function to return the employee ranges dictionary}
\usage{
getEmployeeRanges()
}
\value{
dictionary
}
\description{
Helper function to return the employee ranges dictionary
}
\examples{
getEmployeeRanges()

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
