% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnrollmentsRanges.R
\name{getEnrollmentsRanges}
\alias{getEnrollmentsRanges}
\title{Helper function to return the enrollment range dictionary}
\usage{
getEnrollmentsRanges()
}
\value{
dictionary
}
\description{
Helper function to return the enrollment range dictionary
}
\examples{
getEnrollmentsRanges()

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
