% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEventAppearanceTypes.R
\name{getEventAppearanceTypes}
\alias{getEventAppearanceTypes}
\title{Helper function to return the event appearance types dictionary}
\usage{
getEventAppearanceTypes()
}
\value{
dictionary
}
\description{
Helper function to return the event appearance types dictionary
}
\examples{
getEventAppearanceTypes()

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
