% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEventTypes.R
\name{getEventTypes}
\alias{getEventTypes}
\title{Helper function to return the event types dictionary}
\usage{
getEventTypes()
}
\value{
dictionary
}
\description{
Helper function to return the event types dictionary
}
\examples{
getEventTypes()

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
