% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFundingStages.R
\name{getFundingStages}
\alias{getFundingStages}
\title{Helper function to return the funding stages dictionary}
\usage{
getFundingStages()
}
\value{
dictionary
}
\description{
Helper function to return the funding stages dictionary
}
\examples{
getFundingStages()

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
