% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInvestorStages.R
\name{getInvestorStages}
\alias{getInvestorStages}
\title{Helper function to return the investor stages dictionary}
\usage{
getInvestorStages()
}
\value{
dictionary
}
\description{
Helper function to return the investor stages dictionary
}
\examples{
getInvestorStages()

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
