% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInvestorTypes.R
\name{getInvestorTypes}
\alias{getInvestorTypes}
\title{Helper function to return the investor types dictionary}
\usage{
getInvestorTypes()
}
\value{
dictionary
}
\description{
Helper function to return the investor types dictionary
}
\examples{
getInvestorTypes()

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
