% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIpoStatus.R
\name{getIpoStatus}
\alias{getIpoStatus}
\title{Helper function to return the ipo status dictionary}
\usage{
getIpoStatus()
}
\value{
dictionary
}
\description{
Helper function to return the ipo status dictionary
}
\examples{
getIpoStatus()

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
