% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getJobTypes.R
\name{getJobTypes}
\alias{getJobTypes}
\title{Helper function to return the job types dictionary}
\usage{
getJobTypes()
}
\value{
dictionary
}
\description{
Helper function to return the job types dictionary
}
\examples{
getJobTypes()

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
