% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{getOperators}
\alias{getOperators}
\title{Helper function to return all valid operators}
\usage{
getOperators()
}
\value{
all valid operators
}
\description{
Helper function to return all valid operators
}
\examples{
getOperators()

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
