% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOrganizationCards.R
\name{getOrganizationCards}
\alias{getOrganizationCards}
\title{Helper function to return all possible cards for an organization}
\usage{
getOrganizationCards()
}
\value{
all cards for an organization. Remember to only pick one for the detailed Lookup function.
}
\description{
Helper function to return all possible cards for an organization
}
\examples{
getOrganizationCards()

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
