% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOwnershipTypes.R
\name{getOwnershipTypes}
\alias{getOwnershipTypes}
\title{Helper function to return the ownership types dictionary}
\usage{
getOwnershipTypes()
}
\value{
dictionary
}
\description{
Helper function to return the ownership types dictionary
}
\examples{
getOwnershipTypes()

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
