% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParserForCard.R
\name{getParserForCard}
\alias{getParserForCard}
\title{Helper function to return the correct parsing function for a given card}
\usage{
getParserForCard(card)
}
\arguments{
\item{card}{that is of interest}
}
\value{
the appropriate parser
}
\description{
Helper function to return the correct parsing function for a given card
}
\examples{
getParserForCard("founders")


}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
