% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSchoolMethods.R
\name{getSchoolMethods}
\alias{getSchoolMethods}
\title{Helper function to return the school method dictionary}
\usage{
getSchoolMethods()
}
\value{
dictionary
}
\description{
Helper function to return the school method dictionary
}
\examples{
getSchoolMethods()

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
