% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSchoolPrograms.R
\name{getSchoolPrograms}
\alias{getSchoolPrograms}
\title{Helper function to return the school programs dictionary}
\usage{
getSchoolPrograms()
}
\value{
dictionary
}
\description{
Helper function to return the school programs dictionary
}
\examples{
getSchoolTypes()

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
