% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSchoolTypes.R
\name{getSchoolTypes}
\alias{getSchoolTypes}
\title{Helper function to return the revenue range dictionary}
\usage{
getSchoolTypes()
}
\value{
dictionary
}
\description{
Helper function to return the revenue range dictionary
}
\examples{
getSchoolTypes()

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
