% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listParse.R
\name{listParse}
\alias{listParse}
\title{Helper function to parse a simple list with values (data.frame)}
\usage{
listParse(simpleList)
}
\value{
parsed output
}
\description{
Helper function to parse a simple list with values (data.frame)
}
\examples{
listParse(DataFrameWithValueColumn)


}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
