% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupAcquisitionCard.R
\name{lookupAcquisitionCard}
\alias{lookupAcquisitionCard}
\title{Function to lookup and paginate through a single card for a single acquisition over the Entity Lookup API Endpoint}
\usage{
lookupAcquisitionCard(acquisition_card, acquisition_id)
}
\arguments{
\item{acquisition_card}{card field of interest that will be returned. Only one please!}

\item{acquisition_id}{UUID or permalink of the acquisition you wish to look up. You can provide more than one.}
}
\value{
a data.frame
}
\description{
Takes the entity id to lookup a card and paginate through the entities and parse them into a data.frame
}
\examples{
lookupAcquisitionCard(acquisition_card = "degrees", acquisition_id = "mark-zuckerberg")

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
