% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupAddressCard.R
\name{lookupAddressCard}
\alias{lookupAddressCard}
\title{Function to lookup and paginate through a single card for a single address over the Entity Lookup API Endpoint}
\usage{
lookupAddressCard(address_card, address_id, please_parse = TRUE)
}
\arguments{
\item{address_card}{card field of interest that will be returned. Only one please!}

\item{address_id}{UUID or permalink of the address you wish to look up}

\item{please_parse}{Logical. By default TRUE and will parse your data from a list of data.frames to a final data.frame with empty elements dropping out.}
}
\value{
a data.frame
}
\description{
Takes the entity id to lookup a card and paginate through the entities and parse them into a data.frame
}
\examples{
lookupAddressCard(address_card = "card", address_id = "id")

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
