% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupAddresses.R
\name{lookupAddresses}
\alias{lookupAddresses}
\title{Function to lookup multiple addresses over the Entity Lookup API Endpoint}
\usage{
lookupAddresses(addresses, please_parse = TRUE)
}
\arguments{
\item{please_parse}{TRUE or FALSE. By default TRUE. If set to FALSE, it will return the data directly from the JSON, if set to TRUE, it will parse it into a data.frame object}

\item{persons}{UUID or permalink of the addresses you wish to look up}
}
\value{
either a data.frame (if please_parse = TRUE) or a list of lists (if please_parse = FALSE)
}
\description{
Function to lookup multiple addresses over the Entity Lookup API Endpoint
}
\examples{
lookupAddresses(c("uuid1", "uuid2"))
lookupAddresses(list("uuid1", "uuid2"))

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
