% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupAquisition.R
\name{lookupAquisition}
\alias{lookupAquisition}
\title{Function to lookup a single acquisition over the Entity Lookup API Endpoint}
\usage{
lookupAquisition(id, please_parse = TRUE, print_error = TRUE)
}
\arguments{
\item{id}{UUID or permalink of the acquisition you wish to look up}

\item{please_parse}{Logical. By default TRUE. If set to FALSE, it will return the data directly from the JSON, if set to TRUE, it will parse it into a data.frame object}

\item{print_error}{Logical. Default to TRUE and thus if the request fails, an error message will be printed out.}
}
\value{
either a data.frame (if parse = TRUE) or a list (if parse = FALSE)
}
\description{
Function to lookup a single acquisition over the Entity Lookup API Endpoint
}
\examples{
lookUpAquisition("uuid")

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
