% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupCard.R
\name{lookupCard}
\alias{lookupCard}
\title{Function to lookup a card for a single organization over the Entity Lookup API Endpoint}
\usage{
lookupCard(id, path, card, parser, initial_data)
}
\arguments{
\item{id}{UUID or permalink of the organization you wish to look up}

\item{path}{Type of entity e.g. organizations}

\item{card}{card field of interest that will be returned. Only one please!}

\item{parse}{TRUE or FALSE. By default TRUE. If set to FALSE, it will return the data directly from the JSON, if set to TRUE, it will parse it into a data.frame object}
}
\value{
either a data.frame (if parse = TRUE) or a list (if parse = FALSE)
}
\description{
Takes the entity id to lookup a card and paginate through the entities and parse them into a data.frame
}
\examples{
lookUpDetailForOrganization(id = "facebook", card = "investors")

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
