% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupCategories.R
\name{lookupCategories}
\alias{lookupCategories}
\title{Function to lookup multiple categories over the Entity Lookup API Endpoint}
\usage{
lookupCategories(categories, please_parse = TRUE, print_error = TRUE)
}
\arguments{
\item{please_parse}{Logical. By default TRUE. If set to FALSE, it will return the data directly from the JSON, if set to TRUE, it will parse it into a data.frame object}

\item{print_error}{Logical. Default to TRUE and thus if the request fails, an error message will be printed out.}

\item{persons}{UUID or permalink of the categories you wish to look up}
}
\value{
either a data.frame (if please_parse = TRUE) or a list of lists (if please_parse = FALSE)
}
\description{
Function to lookup multiple categories over the Entity Lookup API Endpoint
}
\examples{
lookUpCategories(c("	f9b14a15-5517-8f38-0562-729ebb54dfdb", "e8804a82-8603-d2e1-c3b5-d241dd8030b9"))
lookUpCategories(list("	f9b14a15-5517-8f38-0562-729ebb54dfdb", "e8804a82-8603-d2e1-c3b5-d241dd8030b9"))

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
