% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupCategoryGroups.R
\name{lookUpCategoryGroups}
\alias{lookUpCategoryGroups}
\title{Function to lookup multiple category groups over the Entity Lookup API Endpoint}
\usage{
lookUpCategoryGroups(category_groups, please_parse = TRUE)
}
\arguments{
\item{please_parse}{TRUE or FALSE. By default TRUE. If set to FALSE, it will return the data directly from the JSON, if set to TRUE, it will parse it into a data.frame object}

\item{persons}{UUID or permalink of the category groups you wish to look up}
}
\value{
either a data.frame (if please_parse = TRUE) or a list of lists (if please_parse = FALSE)
}
\description{
Function to lookup multiple category groups over the Entity Lookup API Endpoint
}
\examples{
lookUpCategoryGroups(c("	f9b14a15-5517-8f38-0562-729ebb54dfdb", "e8804a82-8603-d2e1-c3b5-d241dd8030b9"))
lookUpCategoryGroups(list("	f9b14a15-5517-8f38-0562-729ebb54dfdb", "e8804a82-8603-d2e1-c3b5-d241dd8030b9"))

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
