% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupEntity.R
\name{lookupEntity}
\alias{lookupEntity}
\title{Function to lookup a single entity over the Entity Lookup API Endpoint}
\usage{
lookupEntity(id, path, please_parse = TRUE)
}
\arguments{
\item{id}{UUID or permalink of the person you wish to look up}

\item{path}{The path to the entity e.g. organizations, people, etc.}

\item{please_parse}{TRUE or FALSE. By default TRUE. If set to FALSE, it will return the data directly from the JSON, if set to TRUE, it will parse it into a data.frame object}
}
\value{
either a data.frame (if parse = TRUE) or a list (if parse = FALSE)
}
\description{
Function to lookup a single entity over the Entity Lookup API Endpoint
}
\examples{
lookupEntity("facebook", "organizations", please_parse = T)

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
