% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupEntityCard.R
\name{lookupEntityCard}
\alias{lookupEntityCard}
\title{Function to lookup and paginate through a single card for a single entity over the Entity Lookup API Endpoint}
\usage{
lookupEntityCard(entity_card, entity_id, entity_path, please_parse = TRUE)
}
\arguments{
\item{entity_card}{card field of interest that will be returned. Only one please!}

\item{entity_id}{UUID or permalink of the organization you wish to look up}

\item{entity_path}{Type of entity e.g. organizations}

\item{please_parse}{to pretty parse the final data.frame e.g. removing nestled lists}
}
\value{
a list of paginated data.frames
}
\description{
Takes the entity id to lookup a card and paginate through the entities and parse them into a data.frame
}
\examples{
lookupEntityCard(entity_card = "investors", entity_id = "facebook", entity_path = "organizations")

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
