% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupEventAppearanceCard.R
\name{lookupEventAppearanceCard}
\alias{lookupEventAppearanceCard}
\title{Function to lookup and paginate through a single card for a single event appearance over the Entity Lookup API Endpoint}
\usage{
lookupEventAppearanceCard(
  event_appearance_card,
  event_appearance_id,
  please_parse = TRUE
)
}
\arguments{
\item{event_appearance_card}{card field of interest that will be returned. Only one please!}

\item{event_appearance_id}{UUID or permalink of the event appearance you wish to look up}

\item{please_parse}{Logical. By default TRUE and will parse your data from a list of data.frames to a final data.frame with empty elements dropping out.}
}
\value{
a data.frame
}
\description{
Takes the entity id to lookup a card and paginate through the entities and parse them into a data.frame
}
\examples{
lookupEventAppearanceCard(event_appearance_card = "card", event_appearance_id = "id")

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
