% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupEventAppearances.R
\name{lookupEventAppearances}
\alias{lookupEventAppearances}
\title{Function to lookup multiple event appearances over the Entity Lookup API Endpoint}
\usage{
lookupEventAppearances(event_appearances, please_parse = TRUE)
}
\arguments{
\item{id}{UUID or permalink of the event appearances you wish to look up}

\item{parse}{TRUE or FALSE. By default TRUE. If set to FALSE, it will return the data directly from the JSON, if set to TRUE, it will parse it into a data.frame object}
}
\value{
either a data.frame (if parse = TRUE) or a list (if parse = FALSE)
}
\description{
Function to lookup multiple event appearances over the Entity Lookup API Endpoint
}
\examples{
lookUpEventAppearances(c("uuid1", "uuid2"))

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
