% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupEventCard.R
\name{lookupEventCard}
\alias{lookupEventCard}
\title{Function to lookup and paginate through a single card for a single event over the Entity Lookup API Endpoint}
\usage{
lookupEventCard(event_card, event_id)
}
\arguments{
\item{event_card}{card field of interest that will be returned. Only one please!}

\item{event_id}{UUID or permalink of the event you wish to look up}
}
\value{
a data.frame
}
\description{
Takes the entity id to lookup a card and paginate through the entities and parse them into a data.frame
}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
