% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupEvents.R
\name{lookupEvents}
\alias{lookupEvents}
\title{Function to lookup multiple events over the Entity Lookup API Endpoint}
\usage{
lookupEvents(events, please_parse = TRUE, print_error = TRUE)
}
\arguments{
\item{events}{UUID or permalink of the events you wish to look up}

\item{please_parse}{Logical. By default TRUE. If set to FALSE, it will return the data directly from the JSON, if set to TRUE, it will parse it into a data.frame object}

\item{print_error}{Logical. Default to TRUE and thus if the request fails, an error message will be printed out.}
}
\value{
either a data.frame (if please_parse = TRUE) or a list of lists (if please_parse = FALSE)
}
\description{
Function to lookup multiple events over the Entity Lookup API Endpoint
}
\examples{
lookupEvents(c("6a21f93a-3cae-02ac-44c7-c166c0ed50fa", "135a927a-b238-037e-50e2-d3a2e3511ed2"))
lookupEvents(list("6a21f93a-3cae-02ac-44c7-c166c0ed50fa", "135a927a-b238-037e-50e2-d3a2e3511ed2"))

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
