% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupFundingRounds.R
\name{lookupFundingRounds}
\alias{lookupFundingRounds}
\title{Function to lookup a multiple funding rounds over the Entity Lookup API Endpoint}
\usage{
lookupFundingRounds(funding_rounds, please_parse = TRUE)
}
\arguments{
\item{funding_rounds}{UUID or permalink of the funding rounds you wish to look up}

\item{please_parse}{TRUE or FALSE. By default TRUE. If set to FALSE, it will return the data directly from the JSON, if set to TRUE, it will parse it into a data.frame object}
}
\value{
either a data.frame (if please_parse = TRUE) or a list of lists (if please_parse = FALSE)
}
\description{
Function to lookup a multiple funding rounds over the Entity Lookup API Endpoint
}
\examples{
lookupFundingRounds(c("8a945939-18e0-cc9d-27b9-bf33817b2818", "d950d7a5-79ff-fb93-ca87-13386b0e2feb"))
lookupFundingRounds(list("8a945939-18e0-cc9d-27b9-bf33817b2818", "d950d7a5-79ff-fb93-ca87-13386b0e2feb"))

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
