% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupFunds.R
\name{lookupFunds}
\alias{lookupFunds}
\title{Function to lookup multiple funds over the Entity Lookup API Endpoint}
\usage{
lookupFunds(funds, please_parse = TRUE, print_error = TRUE)
}
\arguments{
\item{please_parse}{Logical. By default TRUE. If set to FALSE, it will return the data directly from the JSON, if set to TRUE, it will parse it into a data.frame object}

\item{print_error}{Logical. Default to TRUE and thus if the request fails, an error message will be printed out.}

\item{id}{UUID or permalink of the funds you wish to look up}
}
\value{
either a data.frame (if parse = TRUE) or a list (if parse = FALSE)
}
\description{
Function to lookup multiple funds over the Entity Lookup API Endpoint
}
\examples{
lookUpFunds(c("uuid1", "uuid2"))

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
