% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupInvestmentCard.R
\name{lookupInvestmentCard}
\alias{lookupInvestmentCard}
\title{Function to lookup and paginate through a single card for a single investment over the Entity Lookup API Endpoint}
\usage{
lookupInvestmentCard(investment_card, investment_id)
}
\arguments{
\item{investment_card}{card field of interest that will be returned. Only one please!}

\item{investment_id}{UUID or permalink of the investment you wish to look up}
}
\value{
a data.frame
}
\description{
Takes the entity id to lookup a card and paginate through the entities and parse them into a data.frame
}
\examples{
lookupInvestmentCard(investment_card = "degrees", investment_id = "mark-zuckerberg")

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
