% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupInvestments.R
\name{lookupInvestments}
\alias{lookupInvestments}
\title{Function to lookup multiple investments over the Entity Lookup API Endpoint}
\usage{
lookupInvestments(investments, please_parse = TRUE)
}
\arguments{
\item{please_parse}{TRUE or FALSE. By default TRUE. If set to FALSE, it will return the data directly from the JSON, if set to TRUE, it will parse it into a data.frame object}

\item{investmentss}{UUID or permalink of the investments you wish to look up}
}
\value{
either a data.frame (if please_parse = TRUE) or a list of lists (if please_parse = FALSE)
}
\description{
Function to lookup multiple investments over the Entity Lookup API Endpoint
}
\examples{
lookupInvestments(c("524986f0-3049-54a4-fa72-f60897a5e61d", "6556ab92-6465-25aa-1ffc-7f8b4b09a476"))
lookupInvestments(list("524986f0-3049-54a4-fa72-f60897a5e61d", "6556ab92-6465-25aa-1ffc-7f8b4b09a476"))

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
