% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupIpos.R
\name{lookupIpos}
\alias{lookupIpos}
\title{Function to lookup multiple ipos over the Entity Lookup API Endpoint}
\usage{
lookupIpos(ipos, please_parse = TRUE)
}
\arguments{
\item{ipos}{UUID or permalink of the ipos you wish to look up}

\item{please_parse}{TRUE or FALSE. By default TRUE. If set to FALSE, it will return the data directly from the JSON, if set to TRUE, it will parse it into a data.frame object}
}
\value{
either a data.frame (if parse = TRUE) or a list (if parse = FALSE)
}
\description{
Function to lookup multiple ipos over the Entity Lookup API Endpoint
}
\examples{
lookUpIpos(c("uuid1", "uuid2"))

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
