% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupJobCard.R
\name{lookupJobCard}
\alias{lookupJobCard}
\title{Function to lookup and paginate through a single card for a single job over the Entity Lookup API Endpoint}
\usage{
lookupJobCard(job_card, job_id, please_parse = TRUE)
}
\arguments{
\item{job_card}{card field of interest that will be returned. Only one please!}

\item{job_id}{UUID or permalink of the job you wish to look up}

\item{please_parse}{Logical. By default TRUE and will parse your data from a list of data.frames to a final data.frame with empty elements dropping out.}
}
\value{
a data.frame
}
\description{
Takes the entity id to lookup a card and paginate through the entities and parse them into a data.frame
}
\examples{
lookupJobCard(job_card = "card", job_id = "id")

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
