% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupJobs.R
\name{lookupJobs}
\alias{lookupJobs}
\title{Function to lookup multiple jobs over the Entity Lookup API Endpoint}
\usage{
lookupJobs(jobs, please_parse = TRUE, print_error = TRUE)
}
\arguments{
\item{please_parse}{Logical. By default TRUE. If set to FALSE, it will return the data directly from the JSON, if set to TRUE, it will parse it into a data.frame object}

\item{print_error}{Logical. Default to TRUE and thus if the request fails, an error message will be printed out.}

\item{persons}{UUID or permalink of the jobs you wish to look up}
}
\value{
either a data.frame (if please_parse = TRUE) or a list of lists (if please_parse = FALSE)
}
\description{
Function to lookup multiple jobs over the Entity Lookup API Endpoint
}
\examples{
lookupJobs(c("uuid1", "uuid2"))
lookupJobs(list("uuid1", "uuid2"))

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
