% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupLocations.R
\name{lookupLocations}
\alias{lookupLocations}
\title{Function to lookup multiple locations over the Entity Lookup API Endpoint}
\usage{
lookupLocations(locations, please_parse = TRUE)
}
\arguments{
\item{please_parse}{TRUE or FALSE. By default TRUE. If set to FALSE, it will return the data directly from the JSON, if set to TRUE, it will parse it into a data.frame object}

\item{persons}{UUID or permalink of the locations you wish to look up}
}
\value{
either a data.frame (if please_parse = TRUE) or a list of lists (if please_parse = FALSE)
}
\description{
Function to lookup multiple locations over the Entity Lookup API Endpoint
}
\examples{
lookupLocations(c("uuid1", "uuid2"))
lookupLocations(list("uuid1", "uuid2"))

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
