% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupOrganizationCard.R
\name{lookupOrganizationCard}
\alias{lookupOrganizationCard}
\title{Function to lookup and paginate through a single card for a single organization over the Entity Lookup API Endpoint}
\usage{
lookupOrganizationCard(organization_card, organization_id, please_parse = TRUE)
}
\arguments{
\item{organization_card}{card field of interest that will be returned. Only one please!}

\item{organization_id}{UUID or permalink of the organization you wish to look up}

\item{please_parse}{Logical. By default TRUE and will parse your data from a list of data.frames to a final data.frame with empty elements dropping out.}
}
\value{
a data.frame
}
\description{
Takes the entity id to lookup a card and paginate through the entities and parse them into a data.frame
}
\examples{
lookupOrganizationCard(organization_card = "investors", organization_id = "facebook")

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
