% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupOrganizations.R
\name{lookUpOrganizations}
\alias{lookUpOrganizations}
\title{Function to lookup a single organization over the Entity Lookup API Endpoint}
\usage{
lookUpOrganizations(organizations, please_parse = TRUE)
}
\arguments{
\item{organizations}{UUID or permalink of the organizations you wish to look up}

\item{please_parse}{TRUE or FALSE. By default TRUE. If set to FALSE, it will return the data directly from the JSON, if set to TRUE, it will parse it into a data.frame object}
}
\value{
either a data.frame (if please_parse = TRUE) or a list of lists (if please_parse = FALSE)
}
\description{
Function to lookup a single organization over the Entity Lookup API Endpoint
}
\examples{
lookUpOrganizations(c("facebook", "google"))
lookUpOrganizations(list("facebook", "google"))

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
