% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupOwnershipCard.R
\name{lookupOwnershipCard}
\alias{lookupOwnershipCard}
\title{Function to lookup and paginate through a single card for a single ownership over the Entity Lookup API Endpoint}
\usage{
lookupOwnershipCard(ownership_card, ownership_id)
}
\arguments{
\item{ownership_card}{card field of interest that will be returned. Only one please!}

\item{ownership_id}{UUID or permalink of the ownership you wish to look up}
}
\value{
a data.frame
}
\description{
Takes the entity id to lookup a card and paginate through the entities and parse them into a data.frame
}
\examples{
lookupOwnershipCard(ownership_card = "card", ownership_id = "id")

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
