% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupOwnerships.R
\name{lookupOwnerships}
\alias{lookupOwnerships}
\title{Function to lookup a multiple ownerships over the Entity Lookup API Endpoint}
\usage{
lookupOwnerships(ownerships, please_parse = TRUE)
}
\arguments{
\item{ownerships}{UUID or permalink of the ownerships you wish to look up}

\item{please_parse}{TRUE or FALSE. By default TRUE. If set to FALSE, it will return the data directly from the JSON, if set to TRUE, it will parse it into a data.frame object}
}
\value{
either a data.frame (if please_parse = TRUE) or a list of lists (if please_parse = FALSE)
}
\description{
Function to lookup a multiple ownerships over the Entity Lookup API Endpoint
}
\examples{
lookupOwnerships(c("3fa85f64-5717-4562-b3fc-2c963f66afa6", "uuid2"))
lookupOwnerships(list("3fa85f64-5717-4562-b3fc-2c963f66afa6", "uuid2"))

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
