% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupPerson.R
\name{lookUpPerson}
\alias{lookUpPerson}
\title{Function to lookup a single person over the Entity Lookup API Endpoint}
\usage{
lookUpPerson(id, please_parse = TRUE)
}
\arguments{
\item{id}{UUID or permalink of the person you wish to look up}

\item{parse}{TRUE or FALSE. By default TRUE. If set to FALSE, it will return the data directly from the JSON, if set to TRUE, it will parse it into a data.frame object}
}
\value{
either a data.frame (if parse = TRUE) or a list (if parse = FALSE)
}
\description{
Function to lookup a single person over the Entity Lookup API Endpoint
}
\examples{
lookUpPerson("facebook")

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
