% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupPersonCard.R
\name{lookupPersonCard}
\alias{lookupPersonCard}
\title{Function to lookup and paginate through a single card for a single person over the Entity Lookup API Endpoint}
\usage{
lookupPersonCard(person_card, person_id)
}
\arguments{
\item{person_card}{card field of interest that will be returned. Only one please!}

\item{person_id}{UUID or permalink of the person you wish to look up}
}
\value{
a data.frame
}
\description{
Takes the entity id to lookup a card and paginate through the entities and parse them into a data.frame
}
\examples{
lookupPersonCard(person_card = "degrees", person_id = "mark-zuckerberg")

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
