% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupPressReferenceCard.R
\name{lookupPressReferenceCard}
\alias{lookupPressReferenceCard}
\title{Function to lookup and paginate through a single card for a single press reference over the Entity Lookup API Endpoint}
\usage{
lookupPressReferenceCard(press_reference_card, press_reference_id)
}
\arguments{
\item{press_reference_card}{card field of interest that will be returned. Only one please!}

\item{press_reference_id}{UUID or permalink of the press reference you wish to look up}
}
\value{
a data.frame
}
\description{
Takes the entity id to lookup a card and paginate through the entities and parse them into a data.frame
}
\examples{
lookupPressReferenceCard(press_reference_card = "card", press_reference_id = "id")

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
