% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupPressReferences.R
\name{lookupPressReferences}
\alias{lookupPressReferences}
\title{Function to lookup multiple press references over the Entity Lookup API Endpoint}
\usage{
lookupPressReferences(press_references, please_parse = TRUE)
}
\arguments{
\item{press_references}{UUID or permalink of the press references you wish to look up}

\item{please_parse}{TRUE or FALSE. By default TRUE. If set to FALSE, it will return the data directly from the JSON, if set to TRUE, it will parse it into a data.frame object}
}
\value{
either a data.frame (if please_parse = TRUE) or a list of lists (if please_parse = FALSE)
}
\description{
Function to lookup multiple press references over the Entity Lookup API Endpoint
}
\examples{
lookupPressReferences(c("uuid1", "uuid2"))
lookupPressReferences(list("uuid1", "uuid2"))

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
