% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make json list function.R
\name{make_json}
\alias{make_json}
\title{Function to construct JSON body request to pass in for the Search API of the Crunchbase API}
\usage{
make_json(
  fields,
  order_by = "",
  sort_direction = "asc",
  query,
  limit = 100L,
  before_id = "",
  after_id = ""
)
}
\arguments{
\item{fields}{Field names which will be returned as columns}

\item{order_by}{Field name by which the response should be ordered}

\item{sort_direction}{Order direction, either "asc" for ascending or "desc" for descending. The default is "asc"}

\item{query}{JSON of query conditions, each can be individuall constructed with the predicate function}

\item{limit}{Number of rows to be returned, default 100, max 2000, min 1}

\item{before_id}{For paginating through responses, use this to go back to the entries before this uuid}

\item{after_id}{For paginating through responses, use this to go forward to the entries after uuid}
}
\value{
an list object that can be passed in as the JSON body parameter in the Search API call once converted to JSON
}
\description{
Function to construct JSON body request to pass in for the Search API of the Crunchbase API
}
\examples{
make_query(fields = "founded_on,locations",
order_by = "founded_on",
sort_direction = "desc",
query = q,
limit = 200)

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
