% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make order function.R
\name{make_order}
\alias{make_order}
\title{Function to use set up the JSON order set within the JSON query body for the Search API endpoint of Crunchbase API}
\usage{
make_order(order_by, sort_direction = "asc")
}
\arguments{
\item{order_by}{Field name by which the search response should be ordered}

\item{sort_direction}{Order direction, either "asc" for ascending or "desc" for descending. The default is "asc"}
}
\value{
an order data.frame that can be passed in as the order parameter in the construct query function to make a Search API call
}
\description{
Function to use set up the JSON order set within the JSON query body for the Search API endpoint of Crunchbase API
}
\examples{
make_order(order_by = "founded_on", sort_direction = "desc")
make_order(order_by = "rank_org")

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
