% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseAddress.R
\name{parseAddress}
\alias{parseAddress}
\title{Function to parse Address Entity from JSON list to Data.frame}
\usage{
parseAddress(data)
}
\arguments{
\item{data}{Returned fields card response in json list form for the entity}
}
\value{
a data.frame of parsed output
}
\description{
Function to parse Address Entity from JSON list to Data.frame
}
\examples{
parseAddress(data$cards$fields)

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
