% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseAquisition.R
\name{parseAquisition}
\alias{parseAquisition}
\title{Function to parse Aquisition Entity from JSON list to Data.frame}
\usage{
parseAquisition(data)
}
\arguments{
\item{data}{Returned fields card response in json list form for the entity}
}
\value{
a data.frame of parsed output
}
\description{
Function to parse Aquisition Entity from JSON list to Data.frame
}
\examples{
parseAquisition(data$cards$fields)


}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
