% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseCategory.R
\name{parseCategory}
\alias{parseCategory}
\title{Function to parse Category Entity from JSON list to Data.frame}
\usage{
parseCategory(data)
}
\arguments{
\item{data}{Returned fields card response in json list form for the entity}
}
\value{
a data.frame of parsed output
}
\description{
Function to parse Category Entity from JSON list to Data.frame
}
\examples{
parseCategory(data$cards$fields)


}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
