% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseCategoryGroup.R
\name{parseCategoryGroup}
\alias{parseCategoryGroup}
\title{Function to parse Category Group Entity from JSON list to Data.frame}
\usage{
parseCategoryGroup(data)
}
\arguments{
\item{data}{Returned fields card response in json list form for the entity}
}
\value{
a data.frame of parsed output
}
\description{
Function to parse Category Group Entity from JSON list to Data.frame
}
\examples{
parseCategoryGroup(data$cards$fields)


}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
