% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseDate.R
\name{parseDate}
\alias{parseDate}
\title{Function to parse Date to the provided level of precision}
\usage{
parseDate(data, var_name)
}
\arguments{
\item{data}{from json}
}
\value{
parsed output
}
\description{
Function to parse Date to the provided level of precision
}
\examples{
parseDateTime(data$cards$fields$someDate)

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
