% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseEvent.R
\name{parseEvent}
\alias{parseEvent}
\title{Function to parse Event Entity from JSON list to Data.frame}
\usage{
parseEvent(data)
}
\arguments{
\item{data}{Returned fields card response in json list form for the entity}
}
\value{
a data.frame of parsed output
}
\description{
Function to parse Event Entity from JSON list to Data.frame
}
\examples{
parseEvent(data$cards$fields)

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
