% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseEventAppearance.R
\name{parseEventAppearance}
\alias{parseEventAppearance}
\title{Function to parse Event Appearance Entity from JSON list to Data.frame}
\usage{
parseEventAppearance(data)
}
\arguments{
\item{data}{Returned fields card response in json list form for the entity}
}
\value{
a data.frame of parsed output
}
\description{
Function to parse Event Appearance Entity from JSON list to Data.frame
}
\examples{
parseInvestment(data$cards$fields)

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
