% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseFundingRound.R
\name{parseFundingRound}
\alias{parseFundingRound}
\title{Function to parse Funding Round Entity from JSON list to Data.frame}
\usage{
parseFundingRound(data)
}
\arguments{
\item{data}{Returned fields card response in json list form for the entity}
}
\value{
a data.frame of parsed output
}
\description{
Function to parse Funding Round Entity from JSON list to Data.frame
}
\examples{
parseFundingRound(data$cards$fields)


}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
