% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseIdentifier.R
\name{parseIdentifier}
\alias{parseIdentifier}
\title{Helper function to parse a identifier data into a dataframe}
\usage{
parseIdentifier(fields_data, field)
}
\arguments{
\item{field}{name that contains the identifier data}
}
\value{
data.frame of uuid, permalink, value, and entity id
}
\description{
Helper function to parse a identifier data into a dataframe
}
\examples{
parseIdentifier("identifier")


}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
