% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseInvestment.R
\name{parseInvestment}
\alias{parseInvestment}
\title{Function to parse Investment Entity from JSON list to Data.frame}
\usage{
parseInvestment(data)
}
\arguments{
\item{data}{Returned fields card response in json list form for the entity}
}
\value{
a data.frame of parsed output
}
\description{
Function to parse Investment Entity from JSON list to Data.frame
}
\examples{
parseInvestment(data$cards$fields)

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
