% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseIpo.R
\name{parseIpo}
\alias{parseIpo}
\title{Function to parse Ipo Entity from JSON list to Data.frame}
\usage{
parseIpo(data)
}
\arguments{
\item{data}{Returned fields card response in json list form for the entity}
}
\value{
a data.frame of parsed output
}
\description{
Function to parse Ipo Entity from JSON list to Data.frame
}
\examples{
parseIpo(data$cards$fields)

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
