% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseJob.R
\name{parseJob}
\alias{parseJob}
\title{Function to parse Job Entity from JSON list to Data.frame}
\usage{
parseJob(data)
}
\arguments{
\item{data}{Returned fields card response in json list form for the entity}
}
\value{
a data.frame of parsed output
}
\description{
Function to parse Job Entity from JSON list to Data.frame
}
\examples{
parseJob(data$cards$fields)

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
