% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseLocation.R
\name{parseLocation}
\alias{parseLocation}
\title{Function to parse Location Entity from JSON list to Data.frame}
\usage{
parseLocation(data)
}
\arguments{
\item{data}{Returned fields card response in json list form for the entity}
}
\value{
a data.frame of parsed output
}
\description{
Function to parse Location Entity from JSON list to Data.frame
}
\examples{
parseLocation(data$cards$fields)


}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
