% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseOrganization.R
\name{parseOrganization}
\alias{parseOrganization}
\title{Function to parse Organization Entity from JSON list to Data.frame}
\usage{
parseOrganization(data)
}
\arguments{
\item{data}{Returned response in json list form for the entity}
}
\value{
a data.frame of parsed output
}
\description{
Function to parse Organization Entity from JSON list to Data.frame
}
\examples{
lookUpOrganization("facebook")

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
