% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseOwnership.R
\name{parseOwnership}
\alias{parseOwnership}
\title{Function to parse Ownership Entity from JSON list to Data.frame}
\usage{
parseOwnership(data)
}
\arguments{
\item{data}{Returned fields card response in json list form for the entity}
}
\value{
a data.frame of parsed output
}
\description{
Function to parse Ownership Entity from JSON list to Data.frame
}
\examples{
parseOwnership(data$cards$fields)

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
