% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsePerson.R
\name{parsePerson}
\alias{parsePerson}
\title{Function to parse Person Entity from JSON list to Data.frame}
\usage{
parsePerson(data)
}
\arguments{
\item{data}{Returned fields card response in json list form for the entity}
}
\value{
a data.frame of parsed output
}
\description{
Function to parse Person Entity from JSON list to Data.frame
}
\examples{
parsePerson(data$cards$fields)

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
