% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsePrincipal.R
\name{parsePrincipal}
\alias{parsePrincipal}
\title{Function to parse Principal Entity from JSON list to Data.frame}
\usage{
parsePrincipal(data)
}
\arguments{
\item{data}{Returned response in json list form for the entity}
}
\value{
a data.frame of parsed output
}
\description{
Function to parse Principal Entity from JSON list to Data.frame
}
\examples{
parsePrincipal(data$fields)

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
