% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printError.R
\name{printError}
\alias{printError}
\title{Helper function to print out the error code along with an explanation of what it means}
\usage{
printError(response_code)
}
\arguments{
\item{status_code}{the response code returned from the http request}
}
\value{
a custom message that is printed to the console
}
\description{
Helper function to print out the error code along with an explanation of what it means
}
\examples{
printError(404)

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
