% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search function.R
\name{search}
\alias{search}
\title{Function to use the Search API endpoint of Crunchbase API}
\usage{
search(path = "organizations", body)
}
\arguments{
\item{path}{What are you searching for? Default: organizations. Other options include:  organizations}

\item{body}{The JSON body that specifies the query, order, limit etc.}
}
\value{
JSON response
}
\description{
Function to use the Search API endpoint of Crunchbase API
}
\examples{
ipos_data <- search(path = "ipos", body = predefinedBody)

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
