% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search entities function.R
\name{search_entities}
\alias{search_entities}
\title{Search for entities}
\usage{
search_entities(path = "organizations", json_list)
}
\description{
Search for entities
}
\examples{
entities_uuids <- search(path = "ipos", json_list = predefinedBody)

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
