% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search entity function.R
\name{search_entity}
\alias{search_entity}
\title{Function to use the Search API endpoint of Crunchbase API}
\usage{
search_entity(path = "organizations", body)
}
\arguments{
\item{path}{What are you searching for? Default: organizations. Other options include:  organizations}

\item{body}{The JSON body that specifies the query, order, limit etc.}
}
\value{
JSON response
}
\description{
Searches for an entity. You need to specify which entity e.g. organization, person, etc. it is and provide a json body. See helper functions make_order and make_query for this.
}
\examples{
ipos_data <- search(path = "ipos", body = predefinedBody)

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
