% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchCondition.R
\name{searchCondition}
\alias{searchCondition}
\title{Function to construct search conditions to be 'rbind'ed into one dataframe to be passed as conditions argument to any search function}
\usage{
searchCondition(subject, verb, object)
}
\arguments{
\item{subject}{the field you are conditioning on e.g. 'location_identifiers'}

\item{verb}{the operator you want to use to connect 'subject' and 'object' e.g. includes, between.. A full list of all operators available can be obtained with the call 'getOperators()'}

\item{object}{the values of the condition as a CHARACTER VECTOR e.g. c('germany', 'france')}
}
\value{
a data.frame row that can be easily 'rbind'ed to other search conditions to be passed in a searchForXXX(search_conditions) call to search for certain entities
}
\description{
Function to construct search conditions to be 'rbind'ed into one dataframe to be passed as conditions argument to any search function
}
\examples{
searchCondition(subject = "location_identifiers", verb = "includes", object = c("germany", "france"))

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
