% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchConditionCurrency.R
\name{searchConditionCurrency}
\alias{searchConditionCurrency}
\title{Function to construct currency type search conditions to be 'rbind'ed into one dataframe to be passed as conditions argument to any search function}
\usage{
searchConditionCurrency(subject, verb, value, currency)
}
\arguments{
\item{subject}{the field you are conditioning on e.g. 'location_identifiers'}

\item{verb}{the operator you want to use to connect 'subject' and 'object' e.g. includes, between.. A full list of all operators available can be obtained with the call 'getOperators()'}

\item{value}{integer or numeric value of the currency condition}

\item{currency}{of the value of the currency condition as a character e.g. "usd"}
}
\value{
a data.frame row that can be easily 'rbind'ed to other search conditions to be passed in a searchForXXX(search_conditions) call to search for certain entities
}
\description{
Function to construct currency type search conditions to be 'rbind'ed into one dataframe to be passed as conditions argument to any search function
}
\examples{
searchConditionCurrency(subject = "money_raised", verb = "gte", value = 10000000L, currency = "usd")

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
