% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchForAddresses.R
\name{searchForAddresses}
\alias{searchForAddresses}
\title{Function to search for addresses that fulfill pre-specified search conditions over the Search API}
\usage{
searchForAddresses(
  search_conditions,
  order_by = "identifier",
  sort_by = "asc",
  result_limit = NA,
  uuids_only = F,
  precise = T
)
}
\arguments{
\item{search_conditions}{the search condition data.frame that the entities have to fulfill}

\item{order_by}{(Optional) field by which to sort the results}

\item{sort_by}{(Optional) sort direction to sort the order field by. Either 'asc' or 'desc'.}

\item{result_limit}{(Optional) numeric limit of the number of results you'd like to have. By default, all results will be returned to you.}

\item{uuids_only}{(Optional) Logical if you want to have only the uuids of the searched for entities returned. By default, after the uuids are retrieved, the information for all entities will be looked up over the entity lookup API endpoint, which takes longer.}
}
\value{
a data.frame
}
\description{
Function to search for addresses that fulfill pre-specified search conditions over the Search API
}
\examples{
searchForAddresses(rbind(searchCondition1,searchCondition2))

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
