% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchForInvestments.R
\name{searchForInvestments}
\alias{searchForInvestments}
\title{Function to search for investments that fulfill pre-specified search conditions over the Search API}
\usage{
searchForInvestments(
  search_conditions,
  order_by = "identifier",
  sort_by = "asc",
  result_limit = NA,
  uuids_only = F,
  precise = T,
  iteration = 1000L
)
}
\arguments{
\item{search_conditions}{the search condition data.frame that the funding_rounds have to fulfill}

\item{order_by}{(Optional) field by which to sort the results}

\item{sort_by}{(Optional) sort direction to sort the order field by. Either 'asc' or 'desc'.}

\item{result_limit}{(Optional) numeric limit of the number of results you'd like to have. By default, all results will be returned to you.}

\item{uuids_only}{(Optional) Logical if you want to have only the uuids of the searched for entities returned. By default, after the uuids are retrieved, the information for all entities will be looked up over the entity lookup API endpoint, which takes longer.}

\item{precise}{(Optional) Logical Default to TRUE. For a large number of results, you may want to set precise to FALSE as realtime updates may make the function otherwise run on indefinitely. When you set precise to FALSE, the last iteration set of results will not be returned. You can modify this parameter with the argument iteration.}

\item{iteration}{(Optional) numeric iteration step. Default to 1000L, max 2000L. The last iteration will be discarded if you set precise to FALSE. Otherwise, this parameter won't affect you.}
}
\value{
a data.frame
}
\description{
Function to search for investments that fulfill pre-specified search conditions over the Search API
}
\examples{
searchForInvestments(rbind(searchCondition1,searchCondition2))

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
