% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchforOrganizations.R
\name{searchforOrganizations}
\alias{searchforOrganizations}
\title{Function to search for organizations that fulfill pre-specified search conditions}
\usage{
searchforOrganizations(conditions, order_by = "value", sort_by = "asc")
}
\arguments{
\item{path}{Type of entity e.g. organizations}

\item{card}{card field of interest that will be returned. Only one please!}

\item{parse}{TRUE or FALSE. By default TRUE. If set to FALSE, it will return the data directly from the JSON, if set to TRUE, it will parse it into a data.frame object}
}
\value{
either a data.frame (if parse = TRUE) or a list (if parse = FALSE)
}
\description{
Function to search for organizations that fulfill pre-specified search conditions
}
\examples{
lookUpDetailForOrganization(id = "facebook", card = "investors")

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
