% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAPIKey.R
\name{setAPIKey}
\alias{setAPIKey}
\title{Function to set API_KEY}
\usage{
setAPIKey()
}
\description{
Function to set API_KEY
}
\examples{
setAPIKey("yourResearchAccessCrunchbaseAPIUserKey")

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
