% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleParse.R
\name{simpleParse}
\alias{simpleParse}
\title{Helper function to parse a simple character or integer list}
\usage{
simpleParse(stringOrNumber)
}
\value{
parsed output
}
\description{
Helper function to parse a simple character or integer list
}
\examples{
simpleParse("3")


}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
