% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocomplete function.R
\name{autocomplete}
\alias{autocomplete}
\title{Function to use the Autocomplete API endpoint of Crunchbase API}
\usage{
autocomplete(
  query,
 
    collection_ids = "organizations,people,funding_rounds,acquisitions,investments,events,press_references,funds,event_appearances,ipos,ownerships,categories,category_groups,locations,jobs",
  limit = 10
)
}
\arguments{
\item{query}{What we are looking for e.g. "Europe" in locations, "Artificial Intelligence" in categories}

\item{collection_ids}{Within which path the matching to the query should be performed e.g. organizations}

\item{limit}{Number of entries to be returned at most (min 1, default 10, max 25)}
}
\value{
the entities matching the query within the specified collection_ids and with the limit specified
}
\description{
Function to use the Autocomplete API endpoint of Crunchbase API
}
\examples{
autocomplete(collection_ids = "categories", query = "Artificial Intelligence", limit = 5)

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
