% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFundingTypes.R
\name{getFundingTypes}
\alias{getFundingTypes}
\title{Helper function to return funding type dictionary}
\usage{
getFundingTypes()
}
\value{
a dictionary of all funding type key and value pairs
}
\description{
Helper function to return funding type dictionary
}
\examples{
getFundingTypes()

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
