% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{getPaths}
\alias{getPaths}
\title{Helper function to return all paths}
\usage{
getPaths()
}
\value{
all valid paths
}
\description{
Helper function to return all paths
}
\examples{
getPaths()

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
