% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupCategoryGroups.R
\name{lookupCategoryGroups}
\alias{lookupCategoryGroups}
\title{Function to lookup multiple category groups over the Entity Lookup API Endpoint}
\usage{
lookupCategoryGroups(category_groups, please_parse = TRUE)
}
\arguments{
\item{please_parse}{Logical. By default TRUE. If set to FALSE, it will return the data directly from the JSON, if set to TRUE, it will parse it into a data.frame object}

\item{persons}{UUID or permalink of the category groups you wish to look up}

\item{print_error}{Logical. Default to TRUE and thus if the request fails, an error message will be printed out.}
}
\value{
either a data.frame (if please_parse = TRUE) or a list of lists (if please_parse = FALSE)
}
\description{
Function to lookup multiple category groups over the Entity Lookup API Endpoint
}
\examples{
lookupCategoryGroups(c("uuid1", "uuid2"))
lookupCategoryGroups(list("uuid1", "uuid2"))

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
