% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupOrganizationCard.R
\name{lookupOrganizationCard}
\alias{lookupOrganizationCard}
\title{Function to lookup and paginate through a single card for a single organization over the Entity Lookup API Endpoint}
\usage{
lookupOrganizationCard(organization_card, organization_id)
}
\arguments{
\item{organization_card}{card field of interest that will be returned. Only one please!}

\item{organization_id}{UUID or permalink of the organization you wish to look up}
}
\value{
a data.frame
}
\description{
Takes the entity id to lookup a card and paginate through the entities and parse them into a data.frame
}
\examples{
lookupOrganizationCard(organization_card = "investors", organization_id = "facebook")

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
