% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchForEntity.R
\name{searchForEntity}
\alias{searchForEntity}
\title{Function to search for some entities matching certain search criteria over the Search API Endpoint}
\usage{
searchForEntity(path, conditions, order_by, sort_by, result_limit, uuids_only)
}
\arguments{
\item{path}{The path to the entity e.g. organizations, people, etc.}

\item{search_conditions}{of what you are looking for}

\item{please_parse}{TRUE or FALSE. By default TRUE. If set to FALSE, it will return the data directly from the JSON, if set to TRUE, it will parse it into a data.frame object}
}
\value{
either a data.frame (if parse = TRUE) or a list (if parse = FALSE)
}
\description{
Function to search for some entities matching certain search criteria over the Search API Endpoint
}
\examples{
lookupEntity("facebook", "organizations", please_parse = T)

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
