% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchForEntity.R
\name{searchForEntity}
\alias{searchForEntity}
\title{Function to search for some entities matching certain search criteria over the Search API Endpoint}
\usage{
searchForEntity(
  path,
  conditions,
  order_by = "identifier",
  sort_by = "asc",
  personal_limit = NA,
  iter = 1000L,
  uuids_only = FALSE
)
}
\arguments{
\item{path}{The path to the entity e.g. organizations, people, etc.}

\item{personal_limit}{Personal set limit of result. Useful when only needing e.g. the top 50 organizations in a certain category.}

\item{iter}{Number of results to be returned per call. Default here 1000L, max 2000L. Usually defaulting to 50.}

\item{search_conditions}{of what you are looking for}

\item{please_parse}{TRUE or FALSE. By default TRUE. If set to FALSE, it will return the data directly from the JSON, if set to TRUE, it will parse it into a data.frame object}
}
\value{
either a data.frame (if parse = TRUE) or a list (if parse = FALSE)
}
\description{
Function to search for some entities matching certain search criteria over the Search API Endpoint
}
\examples{
lookupEntity("facebook", "organizations", please_parse = T)

}
\author{
Layla Rohkohl, \email{byehity@gmail.com}
}
